+++ src/main/java/com/bcncgroup/pricingservice/prices/domain/PriceRepository.java
@@ -3,5 +3,5 @@ package com.bcncgroup.pricingservice.prices.domain;
 import java.time.LocalDateTime;
 
 public interface PriceRepository {
-    Price findPrice(LocalDateTime applicationDate, Integer productId, Integer brandId);
+    Price findPrice(LocalDateTime applicationDate, Long productId, Long brandId);
 }
